requirements.txt
absl-py==0.7.1
astor==0.8.0
audioread==2.1.8
cffi==1.12.3
Cython==0.29.12
decorator==4.4.0
falcon==2.0.0
falcon-multipart==0.2.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
numba==0.45.0
numpy==1.16.4
protobuf==3.9.0
pycparser==2.19
pyworld==0.2.8
pyzmq==18.1.0
resampy==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
SoundFile==0.10.2
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
zmq==0.0.0
